Move get_tracks_including_unlisted
