+++ discovery-provider/src/queries/queries.py
@@ -1,5 +1,6 @@
 import logging # pylint: disable=C0302
 import datetime
+from src.queries.get_tracks_including_unlisted import get_tracks_including_unlisted
 import sqlalchemy
 from sqlalchemy import func, asc, desc, text, case, or_, and_, Integer, Float, Date
 from sqlalchemy.orm import aliased
@@ -23,6 +24,7 @@ from src.queries.query_helpers import get_current_user_id, parse_sort_param, pop
 from src.queries.get_users import get_users
 from src.queries.get_tracks import get_tracks
 from src.queries.get_playlists import get_playlists
+from src.queries.get_tracks_including_unlisted import get_tracks_including_unlisted
 
 logger = logging.getLogger(__name__)
 bp = Blueprint("queries", __name__)
@@ -51,72 +53,9 @@ def get_tracks_route():
 # Expects a JSON body of shape:
 #   { "tracks": [{ "id": number, "url_title": string, "handle": string }]}
 @bp.route("/tracks_including_unlisted", methods=("POST",))
-def get_tracks_including_unlisted():
-    req_data = request.get_json()
-    identifiers = req_data["tracks"]
-    for i in identifiers:
-        helpers.validate_arguments(i, ["handle", "id", "url_title"])
-
-    db = get_db_read_replica()
-    with db.scoped_session() as session:
-        base_query = session.query(Track)
-        filter_cond = []
-
-        # Create filter conditions as a list of `and` clauses
-        for i in identifiers:
-            filter_cond.append(and_(
-                Track.is_current == True,
-                Track.track_id == i["id"]
-            ))
-
-        # Pass array of `and` clauses into an `or` clause as destructured *args
-        base_query = base_query.filter(or_(*filter_cond))
-
-        # Allow filtering of deletes
-        # Note: There is no standard for boolean url parameters, and any value (including 'false')
-        # will be evaluated as true, so an explicit check is made for true
-        if ("filter_deleted" in request.args):
-            filter_deleted = request.args.get("filter_deleted")
-            if (filter_deleted.lower() == 'true'):
-                base_query = base_query.filter(
-                    Track.is_delete == False
-                )
-
-        # Perform the query
-        # TODO: pagination is broken with unlisted tracks
-        query_results = paginate_query(base_query).all()
-        tracks = helpers.query_result_to_list(query_results)
-
-        # Mapping of track_id -> track object from request;
-        # used to check route_id when iterating through identifiers
-        identifiers_map = {track["id"]: track for track in identifiers}
-
-        # If the track is unlisted and the generated route_id does not match the route_id in db,
-        # filter track out from response
-        def filter_fn(track):
-            input_track = identifiers_map[track["track_id"]]
-            route_id = helpers.create_track_route_id(input_track["url_title"], \
-                        input_track["handle"])
-
-            return not track["is_unlisted"] or track["route_id"] == route_id
-
-        tracks = list(filter(filter_fn, tracks))
-
-        if "with_users" in request.args and request.args.get("with_users") != 'false':
-            user_id_list = get_users_ids(tracks)
-            users = get_users_by_id(session, user_id_list)
-            for track in tracks:
-                user = users[track['owner_id']]
-                if user:
-                    track['user'] = user
-
-        track_ids = list(map(lambda track: track["track_id"], tracks))
-
-        # Populate metadata
-        current_user_id = get_current_user_id(required=False)
-        extended_tracks = populate_track_metadata(session, track_ids, tracks, current_user_id)
-
-    return api_helpers.success_response(extended_tracks)
+def get_tracks_including_unlisted_route():
+    tracks = get_tracks_including_unlisted(request.args.to_dict(), request.get_json())
+    return api_helpers.success_response(tracks)
 
 
 @bp.route("/stems/<int:track_id>", methods=("GET",))
